chore: Move native login code inside features
